refactor: Restructure project to clean architecture pattern
- Reorganize packages into data, domain, and presentation layers
- Add data source layer for local and remote data
- Separate repository interfaces and implementations
