+++ app/src/main/java/com/surivalcoding/composerecipeapp/data/mapper/IngredientAmountMapper.kt
@@ -1,8 +1,8 @@
 package com.surivalcoding.composerecipeapp.data.mapper
 
 import com.surivalcoding.composerecipeapp.data.dto.IngredientAmountDto
-import com.surivalcoding.composerecipeapp.data.model.Ingredient
-import com.surivalcoding.composerecipeapp.data.model.IngredientAmount
+import com.surivalcoding.composerecipeapp.domain.model.Ingredient
+import com.surivalcoding.composerecipeapp.domain.model.IngredientAmount
 
 fun IngredientAmountDto.toIngredientAmount() = IngredientAmount(
     ingredient = ingredient?.toIngredient() ?: Ingredient(0, "", ""),
