+++ app/src/main/java/com/surivalcoding/composerecipeapp/data/mapper/IngredientMapper.kt
@@ -1,7 +1,7 @@
 package com.surivalcoding.composerecipeapp.data.mapper
 
 import com.surivalcoding.composerecipeapp.data.dto.IngredientDto
-import com.surivalcoding.composerecipeapp.data.model.Ingredient
+import com.surivalcoding.composerecipeapp.domain.model.Ingredient
 
 fun IngredientDto.toIngredient() = Ingredient(
     id = id ?: 0,
