+++ app/src/main/java/com/surivalcoding/composerecipeapp/data/mapper/RecipeMapper.kt
@@ -1,7 +1,7 @@
 package com.surivalcoding.composerecipeapp.data.mapper
 
 import com.surivalcoding.composerecipeapp.data.dto.RecipeDto
-import com.surivalcoding.composerecipeapp.data.model.Recipe
+import com.surivalcoding.composerecipeapp.domain.model.Recipe
 
 fun RecipeDto.toRecipe() = Recipe(
     category = category ?: "",
