--- app/src/main/java/com/surivalcoding/composerecipeapp/data/repository/RecipeRepository.kt
@@ -1,7 +1,0 @@
-package com.surivalcoding.composerecipeapp.data.repository
-
-import com.surivalcoding.composerecipeapp.data.model.Recipe
-
-interface RecipeRepository {
-    suspend fun getSavedRecipes(): List<Recipe>
-}
