+++ app/src/main/java/com/surivalcoding/composerecipeapp/data/repository/impl/MockRecipeRepositoryImpl.kt
@@ -1,8 +1,9 @@
-package com.surivalcoding.composerecipeapp.data.repository
+package com.surivalcoding.composerecipeapp.data.repository.impl
 
 import com.surivalcoding.composerecipeapp.data.data_source.RecipeDatasource
 import com.surivalcoding.composerecipeapp.data.mapper.toRecipe
-import com.surivalcoding.composerecipeapp.data.model.Recipe
+import com.surivalcoding.composerecipeapp.domain.repository.RecipeRepository
+import com.surivalcoding.composerecipeapp.domain.model.Recipe
 import kotlinx.coroutines.delay
 
 class MockRecipeRepositoryImpl(private val recipeDatasource: RecipeDatasource) : RecipeRepository {
