+++ app/src/main/java/com/surivalcoding/composerecipeapp/domain/model/Ingredient.kt
@@ -1,4 +1,4 @@
-package com.surivalcoding.composerecipeapp.data.model
+package com.surivalcoding.composerecipeapp.domain.model
 
 data class Ingredient(
     val id: Int,
