+++ app/src/main/java/com/surivalcoding/composerecipeapp/domain/model/Recipe.kt
@@ -1,4 +1,4 @@
-package com.surivalcoding.composerecipeapp.data.model
+package com.surivalcoding.composerecipeapp.domain.model
 
 data class Recipe(
     val category: String,
