+++ app/src/main/java/com/surivalcoding/composerecipeapp/domain/repository/RecipeRepository.kt
@@ -1,0 +1,7 @@
+package com.surivalcoding.composerecipeapp.domain.repository
+
+import com.surivalcoding.composerecipeapp.domain.model.Recipe
+
+interface RecipeRepository {
+    suspend fun getSavedRecipes(): List<Recipe>
+}
