+++ app/src/main/java/com/surivalcoding/composerecipeapp/domain/usecase/SearchRecipeUseCase.kt
@@ -1,0 +1,2 @@
+package com.surivalcoding.composerecipeapp.domain.usecase
+
