--- app/src/main/java/com/surivalcoding/composerecipeapp/presentation/saved_recipe/SavedRecipeState.kt
@@ -1,8 +1,0 @@
-package com.surivalcoding.composerecipeapp.presentation.saved_recipe
-
-import com.surivalcoding.composerecipeapp.data.model.Recipe
-
-data class SavedRecipeState(
-    val recipes: List<Recipe> = emptyList(),
-    val isLoading: Boolean = false,
-)
