+++ app/src/main/java/com/surivalcoding/composerecipeapp/presentation/search_recipes/SearchRecipesState.kt
@@ -1,6 +1,6 @@
 package com.surivalcoding.composerecipeapp.presentation.search_recipes
 
-import com.surivalcoding.composerecipeapp.data.model.Recipe
+import com.surivalcoding.composerecipeapp.domain.model.Recipe
 
 data class SearchRecipesState(
     val query : String = "",
